chore(refactor): split code into multiple files
