+++ package.json
@@ -15,6 +15,7 @@
     "gulp-angular-templatecache": "~1.2.1",
     "gulp-concat": "~2.1.7",
     "gulp-header": "~1.0.2",
+    "gulp-footer": "~1.0.5",
     "gulp-jshint": "1.6.4",
     "gulp-minify-css": "~0.3.6",
     "gulp-minify-html": "~0.1.0",
