+++ src/common.css
@@ -1,3 +1,11 @@
+/*!
+ * ui-select
+ * http://github.com/angular-ui/ui-select
+ * Version: 0.9.5 - 2014-12-12T16:07:20.859Z
+ * License: MIT
+ */
+
+
 /* Style when highlighting a search. */
 .ui-select-highlight {
   font-weight: bold;
@@ -74,14 +82,11 @@
 
 .ui-select-bootstrap .ui-select-toggle {
   position: relative;
-
-  /* Instead of center because of .btn */
-  text-align: left;
 }
 
 .ui-select-bootstrap .ui-select-toggle > .caret {
   position: absolute;
-  height: 10px;
+  height: 100%;
   top: 50%;
   right: 10px;
   margin-top: -2px;
@@ -97,6 +102,11 @@
   border-radius: 4px; /* FIXME hardcoded value :-/ */
   border-top-right-radius: 0;
   border-bottom-right-radius: 0;
+}
+
+.ui-select-bootstrap > .ui-select-match {
+  /* Instead of center because of .btn */
+  text-align: left;
 }
 
 .ui-select-bootstrap > .ui-select-match > .caret {
